requirements.txt
requests~=2.32.3
python-dotenv~=1.1.0
tqdm~=4.66.1